requirements.txt
flask
requests
pandas
numpy
pytz
pulp
thefuzz
unidecode
mlb-statsapi
nba-api
gunicorn